OnlyPaws/requirements.txt
pyTelegramBotAPI
psycopg2-binary
huggingface_hub
matplotlib
bcrypt
requests
python-dotenv
